[core] Attribute: Add a `flattenedChildren` property
`flattenedChildren` returns the list of all the attributes that refer
to this attribute as their parent (either direct or indirect) through
the `root` property. The search for the children attributes is recursive
and alllows to retrieve at once all the nested attributes, independently
from their level.

At the moment, only `ListAttribute` and `GroupAttribute` will return a
non-empty list of flattened children attributes.