[core] Attribute: Parent's `exposed` property takes precedence over description's
The `exposed` property, which determines whether the attribute is
displayed on the upper part of the node in the Graph Editor, is set
for each attribute individually in their node's description.

If an attribute has a parent (meaning it depends on a `GroupAttribute` or
a `ListAttribute`) whose `exposed` property value differs, it does not
make sense to display it separately from it. The attribute's `exposed`
should align with its parent's.